Move ArgumentUtil class
